Backend/requirements.txt
# Environment and Configuration
python-dotenv==1.0.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20

# Data Models and Validation
pydantic==2.11.7

# AI and Language Processing
openai==1.99.7
langfuse

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
PyPDF2==3.0.1

# Document Processing
python-docx==1.1.0
reportlab==4.4.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Cloud Services
firebase-admin==6.5.0
google-cloud-storage==2.14.0
